tp_iw/requirements.txt
asgiref==3.2.10
backcall==0.2.0
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
decorator==4.4.2
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1
django-allauth==0.42.0
django-ckeditor==6.0.0
django-extensions==3.0.8
django-heroku==0.3.1
django-js-asset==1.2.2
graphviz==0.14.1
gunicorn==20.0.4
idna==2.10
ipython==7.17.0
ipython-genutils==0.2.0
jedi==0.17.2
oauthlib==3.1.0
parso==0.7.1
pickleshare==0.7.5
Pillow==7.2.0
prompt-toolkit==3.0.6
psycopg2==2.8.5
pydot==1.4.1
pydot2==1.0.33
pydotplus==2.0.2
Pygments==2.6.1
pyparsing==2.4.7
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
traitlets==4.3.3
urllib3==1.25.10
wcwidth==0.2.5
whitenoise==5.2.0
# django-haystack  # cuando hagan el release que anda con django 3.0.
# Mientras:
git+https://github.com/django-haystack/django-haystack.git@08dd9471a0137ba87b8145cf1ec5ff9d8d498a2e
elasticsearch
whoosh
